[MOVE/#153] UseCase 패키징 설정
