+++ domain/src/main/kotlin/co/orange/domain/usecase/auth/AuthChangeTokenAndSaveUseCase.kt
@@ -1,4 +1,4 @@
-package co.orange.domain.usecase
+package co.orange.domain.usecase.auth
 
 import co.orange.domain.entity.request.AuthRequestModel
 import co.orange.domain.repository.AuthRepository
