+++ domain/src/main/kotlin/co/orange/domain/usecase/buy/BuyGetOrderInfoUseCase.kt
@@ -1,4 +1,4 @@
-package co.orange.domain.usecase
+package co.orange.domain.usecase.buy
 
 import co.orange.domain.repository.BuyRepository
 import javax.inject.Inject
